Use relationships to connect Results, Attacks, and Teams together.
Previously, only some links used relationships while others used hand-written SQL statements to query linked objects.
Trying to link these together with relationships results in fun circular dependency issues, which I'm guessing is why the raw sql statements were used.
This partly gets around the circular dependency bs by combining the Attack and Result classes into a single file, I might undo this later if I can find
another solution.
